+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel, support table now known as lookup table.
 - NullableCheckBox, FlatLaf hack removed.
+- QueryInspector added, available with Ctrl-Alt-Q in EntityTablePanel, when using a local connection.
 
 ## 0.18.34
 ### is.codion.common.core